Move EM voters screen and stats screen into separate files
